+++ gopacttest/self_bootstrap_release_gate.go
@@ -544,6 +544,7 @@ func selfBootstrapFeatureCoverageCommands() []string {
 		"go test -count=1 ./a2a ./gopacttest/a2aconformance",
 		"go test -count=1 -run ExampleNewHTTPRegistryHandler ./a2a",
 		"go test -count=1 ./a2a -run TestMeshSyncEvery",
+		"go test -count=1 ./a2a -run TestMeshSyncEnvEvery",
 		"go test -count=1 ./cmd/gopact",
 		"go test -count=1 -run Channel . ./gopacttest",
 		"go test -count=1 . ./sandbox ./gopacttest/secretconformance ./gopacttest/promptinjectionconformance",